# docs/swagger.yaml
basePath: /
definitions:
  check-id-api_api_handlers_onboarding.Onboarding:
    properties:
      method:
        type: string
      url:
        type: string
    type: object
...
  onboarding.RequestProcessOnboarding:
    properties:
      document_back:
        type: string
      document_front:
        type: string
      onboarding:
        type: string
      selfie:
        type: string
      user_id:
        type: string
    type: object
  onboarding.RequestValidationIdentity:
    properties:
      face_image:
        type: string
      user_id:
        type: string
      validation_id:
        type: integer
    type: object
  onboarding.ResProcessOnboarding:
    properties:
      code:
        type: integer
      data: {}
      error:
        type: boolean
      msg:
        type: string
      type:
        type: integer
    type: object
  onboarding.resCreateOnboarding:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/check-id-api_api_handlers_onboarding.Onboarding'
      error:
        type: boolean
      msg:
        type: string
      type:
        type: integer
    type: object
...
host: 172.147.77.149:50050
info:
  contact:
    email: luis.lucero@bjungle.net
    name: API Support
  description: Api para OnBoarding y validación de identidad
  license:
    name: Software Owner
    url: https://www.bjungle.net/terms/licenses
  termsOfService: https://www.bjungle.net/terms/
  title: Check ID OnBoarding
  version: "1.4"
paths:
...
  /api/v1/onboarding/:
    post:
      consumes:
      - application/json
      description: Método que permite iniciar el enrolamiento de un usuario que puede
        ser desde un tercero o desde el mismo sistema
      parameters:
      - default: Bearer <Add access token here>
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: Datos para el enrolamiento del usuario
        in: body
        name: resCreateOnboarding
        required: true
        schema:
          $ref: '#/definitions/onboarding.resCreateOnboarding'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/onboarding.resCreateOnboarding'
      summary: Método que permite iniciar el enrolamiento de un usuario
      tags:
      - Onboarding
  /api/v1/onboarding/process:
    post:
      consumes:
      - application/json
      description: Método que permite terminar el enrolamiento de un usuario que ha
        sido validado desde OnlyOne
      parameters:
      - default: Bearer <Add access token here>
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: Datos para validar el enrolamiento del usuario
        in: body
        name: RequestProcessOnboarding
        required: true
        schema:
          $ref: '#/definitions/onboarding.RequestProcessOnboarding'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/onboarding.ResProcessOnboarding'
      summary: Método que permite terminar el enrolamiento de un usuario
      tags:
      - Onboarding
  /api/v1/onboarding/validate-identity:
    post:
      consumes:
      - application/json
      description: Método que permite finalizar la validación de identidad de un usuario
        por la aplicación de OnlyOne
      parameters:
      - default: Bearer <Add access token here>
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: Datos para validar la identidad del usuario
        in: body
        name: RequestValidationIdentity
        required: true
        schema:
          $ref: '#/definitions/onboarding.RequestValidationIdentity'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/onboarding.ResProcessOnboarding'
      summary: Método que permite finalizar la validación de identidad de un usuario
      tags:
      - Onboarding
...
swagger: "2.0"
...